// src/components/atoms/Typography/Typography.types.ts
export type TypographyVariant =
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "title1"
  | "title2"
  | "title3"
  | "body1"
  | "body2"
  | "body3"
  | "label1"
  | "label2"
  | "label3"
  | "label4"
  | "caption1"
  | "caption2"
  | "btn-sm"
  | "btn-md"
  | "btn-lg";

export type TypographyColor =
  | "text-primary"
  | "text-secondary"
  | "text-disabled"
  | "text-placeholder"
  | "text-link"
  | "text-white"
  | "primary"
  | "info"
  | "success"
  | "warning"
  | "danger";